Add: checkboxのチェックの有無で挙動変更されるロジック導入、jsファイル分けた
